[refactor] 파라미터 재정의 및 쿼리 재작성
- Object Mapper 파라미터 재정의
- DB Schema 변경점에 따른 쿼리 재작성
