--- src/main/java/com/suitecare/suitecare/api/mate/mapper/LocationMapper.java
@@ -1,12 +1,0 @@
-package com.suitecare.suitecare.api.mate.mapper;
-
-import com.suitecare.suitecare.api.mate.dto.LocationDTO;
-import org.apache.ibatis.annotations.Mapper;
-
-import java.util.List;
-
-@Mapper
-public interface LocationMapper {
-    List<LocationDTO> findLocationById(String id);
-    void createLocation(Integer resume_id, String[] location);
-}
