+++ src/main/java/com/suitecare/suitecare/api/mate_resume/mapper/CertificateMapper.java
@@ -1,0 +1,12 @@
+package com.suitecare.suitecare.api.mate_resume.mapper;
+
+import com.suitecare.suitecare.api.mate_resume.dto.CertificateDTO;
+import org.apache.ibatis.annotations.Mapper;
+
+import java.util.List;
+
+@Mapper
+public interface CertificateMapper {
+    List<CertificateDTO> findCertificateById(String id);
+    void createCertificate(String mateResumeId, List<CertificateDTO> certificateList);
+}
