+++ src/main/java/com/suitecare/suitecare/api/mate_resume/mapper/MainServiceMapper.java
@@ -1,0 +1,12 @@
+package com.suitecare.suitecare.api.mate_resume.mapper;
+
+import com.suitecare.suitecare.api.mate_resume.dto.MainServiceDTO;
+import org.apache.ibatis.annotations.Mapper;
+
+import java.util.List;
+
+@Mapper
+public interface MainServiceMapper {
+    List<MainServiceDTO> findMainServiceById(String id);
+    void createMainService(String mateResumeId, List<MainServiceDTO> mainServiceList);
+}
