+++ src/main/resources/mapper/CareerMapper.xml
@@ -2,22 +2,21 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.suitecare.suitecare.api.mate.mapper.CareerMapper">
-
-    <select id="findCareerById" resultType="CareerDTO">
-        SELECT career.id, name, job_name, date_start, date_end
-        FROM career
-        JOIN materesume m ON m.id = mate_resume_id
-        WHERE m.member_id =#{id};
-    </select>
+<mapper namespace="com.suitecare.suitecare.api.mate_resume.mapper.CareerMapper">
 
     <insert id="createCareer">
         INSERT INTO career(mate_resume_id, name, job_name, date_start, date_end)
         VALUES
-        <foreach collection="career" item="item" separator=", ">
-            (#{resume_id}, #{item.name}, #{item.job_name}, #{item.date_start}, #{item.date_end})
+        <foreach collection="careerList" item="career" open="(" close=")" separator=",">
+            #{mateResumeId}, #{career.name}, #{career.jobName}, #{career.dateStart}, #{career.dateEnd}
         </foreach>
     </insert>
 
+    <select id="findCareerById" resultType="CareerDTO">
+        SELECT career.id, name, job_name as jobName, date_start as dateStart, date_end as dateEnd
+        FROM career
+        JOIN materesume m ON m.id = mate_resume_id
+        WHERE m.id =#{id};
+    </select>
 
 </mapper>
