+++ src/main/resources/mapper/LocationMapper.xml
@@ -2,20 +2,21 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.suitecare.suitecare.api.mate.mapper.LocationMapper">
+<mapper namespace="com.suitecare.suitecare.api.mate_resume.mapper.LocationMapper">
+
+    <insert id="createLocation">
+        INSERT INTO location(mate_resume_id, name)
+        VALUES
+        <foreach collection="locationList" item="location" open="(" close=")" separator=",">
+            #{mateResumeId}, #{location.name}
+        </foreach>
+    </insert>
 
     <select id="findLocationById" resultType="LocationDTO">
         SELECT l.id, name
         FROM location l
         JOIN materesume m ON m.id = l.mate_resume_id
-        WHERE m.member_id = #{id};
+        WHERE m.id = #{id};
     </select>
 
-    <insert id="createLocation">
-        INSERT INTO location(mate_resume_id, name)
-        VALUES
-        <foreach collection="location" index="idx" item="loc" separator=", ">
-            (#{resume_id}, #{loc})
-        </foreach>
-    </insert>
 </mapper>
