+++ src/main/resources/mapper/MainServiceMapper.xml
@@ -2,22 +2,21 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.suitecare.suitecare.api.mate.mapper.MainServiceMapper">
+<mapper namespace="com.suitecare.suitecare.api.mate_resume.mapper.MainServiceMapper">
+
+    <insert id="createMainService">
+        INSERT INTO mainService(mate_resume_id, name)
+        VALUES
+        <foreach collection="mainServiceList" item="mainService" open="(" close=")" separator=",">
+            #{mateResumeId}, #{mainService.name}
+        </foreach>
+    </insert>
 
     <select id="findMainServiceById" resultType="MainServiceDTO">
         SELECT s.id, name
         FROM mainservice s
         JOIN materesume m ON m.id = s.mate_resume_id
-        WHERE m.member_id = #{id};
+        WHERE m.id = #{id};
     </select>
-
-    <insert id="createMainService">
-        INSERT INTO mainservice(mate_resume_id, name)
-        VALUES
-        <foreach collection="mainService" item="service" separator=", ">
-            (#{resume_id}, #{service})
-        </foreach>
-    </insert>
-
 
 </mapper>
